refact: S3 -> GCS로 예외 코드 명칭 변경
