+++ src/main/java/umc/cockple/demo/domain/file/exception/GcsException.java
@@ -3,7 +3,7 @@ package umc.cockple.demo.domain.file.exception;
 import umc.cockple.demo.global.exception.GeneralException;
 import umc.cockple.demo.global.response.code.BaseErrorCode;
 
-public class S3Exception extends GeneralException {
+public class GcsException extends GeneralException {
 
-    public S3Exception(BaseErrorCode code) { super(code); }
+    public GcsException(BaseErrorCode code) { super(code); }
 }
